# docs/openapi/mdm-openapi1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MDM Device Posture API",
    "version": "1.1.0",
    "description": "Canonical API contract for MDM  device posture system. This JSON spec covers agent ingestion, normalization, evaluation, scoring, decisioning, remediation, policy management, catalog, lookups, and admin jobs."
  },
  "servers": [
    {
      "url": "https://mdm.24onlinebilling.com",
      "description": "Production"
    },
    {
      "url": "https://staging-mdm.24onlinebilling.com",
      "description": "Staging"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/catalog/applications": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "List application catalog",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "in": "query",
            "name": "os_type",
            "schema": {
              "$ref": "#/components/schemas/OsType"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Catalog list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "Create catalog entry",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationCatalogUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationCatalogEntry"
                }
              }
            }
          }
        }
      }
    },
    "/v1/catalog/applications/{application_catalog_id}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Get catalog entry",
        "parameters": [
          {
            "$ref": "#/components/parameters/ApplicationCatalogId"
          }
        ],
        "responses": {
          "200": {
            "description": "Catalog entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationCatalogEntry"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Catalog"
        ],
        "summary": "Patch catalog entry",
        "parameters": [
          {
            "$ref": "#/components/parameters/ApplicationCatalogId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationCatalogUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationCatalogEntry"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "Page": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
      "Size": {
        "in": "query",
        "name": "size",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 20
        }
      },
...
      "ApplicationCatalogId": {
        "in": "path",
        "name": "application_catalog_id",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
    },
...
    "schemas": {
...
      "PagedEnvelope": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "page": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "size": {
                "type": "integer"
              },
              "total_elements": {
                "type": "integer"
              },
              "total_pages": {
                "type": "integer"
              }
            }
          }
        }
      },
      "OsType": {
        "type": "string",
        "enum": [
          "ANDROID",
          "IOS",
          "WINDOWS",
          "MACOS",
          "LINUX"
        ]
      },
...
      "ApplicationCatalogEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "os_type": {
            "$ref": "#/components/schemas/OsType"
          },
          "package_id": {
            "type": "string",
            "nullable": true
          },
          "app_name": {
            "type": "string"
          },
          "publisher": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ApplicationCatalogUpsertRequest": {
        "type": "object",
        "required": [
          "os_type",
          "app_name"
        ],
        "properties": {
          "os_type": {
            "$ref": "#/components/schemas/OsType"
          },
          "package_id": {
            "type": "string"
          },
          "app_name": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          }
        }
      },
...
    }
  }
}